refactor: extract hook `useDialog`
